# src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.0'
  title: Care Platform
  description: 'This is the API definition for the Care Platform.

  Authentication  Nodes:
  Click Authorize to enter User and Component authorization codes.

  - userAuth: The JWT token provided by the login command. Since the swagger 2.0 (OpenAPI) specification does
       not natively support bearer authentication. Please use this APIKey field. Prefix it with the word Bearer,
       like this: Bearer JWT_TOKEN
  - componentAuth: The key of the component, this can be obtained by making GET requests to .../components.'
host: localhost:10010
basePath: /v0
schemes:
- http
#s
consumes:
- application/json
produces:
- application/json
securityDefinitions:
   userAuth:
     type: apiKey
     in: header
     name: X-API-Key

   componentAuth:
     type: apiKey
     in: header
     name: X-API-Key
paths:
...
  /clients/{clientID}/components:
    x-swagger-router-controller: components
    post:
      security:
      - userAuth: []
      description: Create a new component for this client, generating a key to be given to the component. 
      summary:  Register a new component
      tags:
      - client
      operationId: newComponent
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: ID of client to which the component should be added
      - name: body
        in: body
        required: true
        description: New component information
        schema:
          $ref: '#/definitions/NewComponent'
      responses:
        201:
          description: The component was created
          schema:
            $ref: '#/definitions/Component'
        401:
          description: Invalid credentials supplied
        404:
          description: The client does not exist
          
    get:
      security:
      - userAuth: []
      description: See the components associated with this client 
      summary:  Get client's components
      tags:
      - client
      operationId: getComponents
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Component'
        401:
          description: Invalid credentials supplied
        404:
          description: The client does not exist
  /clients/{clientID}/components/{componentID}:
    x-swagger-router-controller: components
    get:
      security:
      - userAuth: []
      description: See a specific component of a client
      summary:  Get a single component of the client
      tags:
      - client
      operationId: getComponentById
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      responses:
        200:
          description: OK
          schema:
           $ref: '#/definitions/Component'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist
          
    delete:
      security:
      - userAuth: []
      description: Delete a component of a client
      summary:  Delete a single component of the client
      tags:
      - client
      operationId: deleteComponentById
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to delete
      responses:
        200:
          description: OK
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist
  /clients/{clientID}/components/{componentID}/permissions:
    x-swagger-router-controller: permissions
    get:
      security:
      - userAuth: []
      description: If this compoent has any accepted permssions, you can get them here.
      summary:  See the permissions the component has been granted.
      tags:
      - client
      operationId: getPermissions
      produces:
      - application/json
      parameters:
      - name: clientID
        in: path
        required: true
        type: integer
        description: Get the components of this client
      - name: componentID
        in: path
        required: true
        type: integer
        description: The component to get
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Permission'
        401:
          description: Invalid credentials supplied
        404:
          description: The client or component does not exist, or you are not the client's admin.
...
  /clients/{clientID}/events:
      x-swagger-router-controller: events
      get:
        security:
        - userAuth: []
        description: Get events for all components of this client.
        summary:  Get all events
        tags:
        - client
        operationId: getAllEvents
        produces:
        - application/json
        parameters:
        - name: clientID
          in: path
          required: true
          type: integer
          description: Get the components of this client
        responses:
          200:
            description: Events
            schema:
              $ref: '#/definitions/EventsList'
          401:
            description: Invalid credentials supplied
          404:
            description: The client does not exist or has no components
  /clients/{clientID}/components/{componentID}/events:
      x-swagger-router-controller: events
      get:
        security:
        - userAuth: []
        description: Get events from this component
        summary:  Get events from this component
        tags:
        - client
        operationId: getComponentEvents
        produces:
        - application/json
        parameters:
        - name: clientID
          in: path
          required: true
          type: integer
          description: Get the components of this client
        - name: componentID
          in: path
          required: true
          type: integer
          description: Get events from this component
        responses:
          200:
            description: Events
            schema:
              $ref: '#/definitions/EventsList'
          401:
            description: Invalid credentials supplied
          404:
            description: The client or component does not exist
...
  /components/events:
    x-swagger-router-controller: events
    post:
      security:
      - componentAuth: []
      description: 'You should include information on the device that has emitted it, along with its event type and
      any extra details'
      summary:  Send an information on a new event
      tags:
      - component
      operationId: postEvent
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: Event
        schema:
          type: object
          required:
          - type
          - deviceInstance
          - details
          properties:
            type:
              type: string
              example: 'Opened'
            details:
              type: string
              example: 'The pill box was opened'
            deviceInstance:
              $ref: '#/definitions/DeviceInstance'
      responses:
        200:
          description: The event was recieved correctly
        401:
          description: 'The component key you supplied is invalid, or the user has not granted the required
          permissions to post this event.'
definitions:
...
  NewComponent:
    title: Component to create
    type: object
    properties:
      type:
        type: string
        enum: 
        - event
        - model
        example: event
      name: 
        type: string
        example: Pill Box
      apiEndPoint:
        type: string
        example: pillboxexample.com/api
    required:
    - name
    - type
  Component:
    title: Component
    type: object
    allOf:
    - $ref: '#/definitions/NewComponent'
    properties:
      id:
        type: integer
        example: 0
      key:
        type: string
        example: pJN2GqumRbOmtNuCgKlhxjDgRWepazFGuCiikGcizc3GtUM9d8MXz5LCfEI0DjS0HzeiWODd9VGCjcTenttVnsCxG4RH1Sj1QMHe
    required:
    - id
    - key
    - type
  Permission:
    type: object
    title: Component Permission
    properties:
      type:
        type: string
        enum:
        - event
        - device
        - alert
        example: 'device'
      name:
        type: string
        example: "Pill Box"
    required:
    - type
    - name
...
  DeviceInstance:
    type: object
    title: Device Instance
    properties:
      type:
        type: string
        example: Pill Box
      properties:
        type: string
        example: "{\"location\": \"Kitchen\"}"
    required:
    - type
    - properties

  #Events definitions
  EventsList:
    type: array
    items:
      type: object
      required:
      - type
      - deviceInstance
      - details
      properties:
        component:
          type: string
          example: 'Pill Box Component'
        date:
          type: string
          example: '2018-02-27T18:42:29.000Z'
        type:
          type: string
          example: 'Opened'
        details:
          type: string
          example: 'The pill box was opened'
        deviceInstance:
          $ref: '#/definitions/DeviceInstance'
        
    #other definitions
...